openvm_stark_backend
Module gkr

GKR batch prover for Grand Product and LogUp lookup arguments.

Structs

GkrArtifact

Values of interest obtained from the execution of the GKR protocol.

GkrBatchProof

Batch GKR proof.

GkrMask

Stores two evaluations of each column in a GKR layer.

InvalidNumMaskColumnsError

Error indicating the mask has an invalid number of columns for a gate’s operation.

NotConstantPolyError

Error returned when a polynomial is expected to be constant but it is not.

NotOutputLayerError

Enums

Gate

Defines how a circuit operates locally on two input rows to produce a single output row. This local 2-to-1 constraint is what gives the whole circuit its “binary tree” structure.

GkrError

Error encountered during GKR protocol verification.

Layer

Represents a layer in a binary tree structured GKR circuit.

Functions

correct_sum_as_poly_in_first_variable

Computes r(t) = sum_x eq((t, x), y[-k:]) * p(t, x) from evaluations of f(t) = sum_x eq(({0}^(n - k), 0, x), y) * p(t, x).

partially_verify_batch

Partially verifies a batch GKR proof.

prove_batch

Batch proves lookup circuits with GKR.

Type Aliases

LayerIndex

GKR layer index where 0 corresponds to the output layer.