openvm_stark_backend::gkr
Function partially_verify_batch

pub fn partially_verify_batch<F: Field>(
    gate_by_instance: Vec<Gate>,
    proof: &GkrBatchProof<F>,
    challenger: &mut impl FieldChallenger<F>,
) -> Result<GkrArtifact<F>, GkrError<F>>

Partially verifies a batch GKR proof.

On successful verification the function returns a GkrArtifact which stores the out-of-domain point and claimed evaluations in the input layer columns for each instance at the OOD point. These claimed evaluations are not checked in this function - hence partial verification.