openvm_stark_backend
Module interaction

Log-up permutation argument implementation as RAP.

Modules

debug

Interaction debugging tools

rap

An AIR with specified interactions can be augmented into a RAP. This module auto-converts any [Air] implemented on an InteractionBuilder into a Rap.

stark_log_up

trace

Structs

Interaction

RapPhaseProverData

RapPhaseShape

RapPhaseVerifierData

Enums

InteractionType

RapPhaseSeqKind

Supported challenge phases in a RAP.

Traits

HasInteractionChunkSize

InteractionBuilder

An [AirBuilder] with additional functionality to build special logUp arguments for communication between AIRs across buses. These arguments use randomness to add additional trace columns (in the extension field) and constraints to the AIR.

RapPhaseSeq

Defines a particular protocol for the “after challenge” phase in a RAP.

Type Aliases

SymbolicInteraction