openvm_stark_backend
Module rap

Trait for RAP (Randomized AIR with Preprocessing)

RAP (Randomized Air with Preprocessing)

See https://hackmd.io/@aztec-network/plonk-arithmetiization-air for formal definition.

Traits

AnyRap

RAP trait for all-purpose dynamic dispatch use. This trait is auto-implemented if you implement Air and BaseAirWithPublicValues and PartitionedBaseAir traits.

BaseAirWithPublicValues

An AIR with 0 or more public values. This trait will be merged into Plonky3 in PR: https://github.com/Plonky3/Plonky3/pull/470

PartitionedBaseAir

An AIR with 1 or more main trace partitions.

PermutationAirBuilderWithExposedValues

Permutation AIR builder that exposes certain values to both prover and verifier after the permutation challenges are drawn. These can be thought of as “public values” known after the challenges are drawn.

Rap

An AIR that works with a particular AirBuilder which allows preprocessing and injected randomness.

Functions

get_air_name

Automatically derives the AIR name from the type name for pretty display purposes.