openvm_stark_backend
Module sumcheck

Sum-check protocol that proves and verifies claims about sum_x g(x) for all x in {0, 1}^n.

MultivariatePolyOracle provides methods for evaluating sums and making transformations on g in the context of the protocol. It is intended to be used in conjunction with prove_batch to generate proofs.

Structs

SumcheckArtifacts

SumcheckProof

Enums

SumcheckError

Sum-check protocol verification error.

Constants

MAX_DEGREE

Max degree of polynomials the verifier accepts in each round of the protocol.

Functions

partially_verify

Partially verifies a sum-check proof.

prove_batch

Performs sum-check on a random linear combinations of multiple multivariate polynomials.

Type Aliases

RoundIndex

Sum-check round index where 0 corresponds to the first round.