openvm_stark_backend::sumcheck
Function partially_verify

pub fn partially_verify<F: Field>(
    claim: F,
    proof: &SumcheckProof<F>,
    challenger: &mut impl FieldChallenger<F>,
) -> Result<(Vec<F>, F), SumcheckError<F>>

Partially verifies a sum-check proof.

Only “partial” since it does not fully verify the prover’s claimed evaluation on the variable assignment but checks if the sum of the round polynomials evaluated on 0 and 1 matches the claim for each round. If the proof passes these checks, the variable assignment and the prover’s claimed evaluation are returned for the caller to validate otherwise an Err is returned.

Output is of the form (variable_assignment, claimed_eval).